+++ NEWS.md
@@ -1,3 +1,9 @@
+# inlmisc 0.2.6
+
+- Add `endian` argument to `ReadModflowBinary` function.
+  Argument describes the endian-ness (or byte-order) of the binary file and is required for calls to the `readBin` function.
+  Thanks to Professor Brian Ripley for identifying this issue.
+
 # inlmisc 0.2.5
 
 - In `SummariseBudget` function, the `desc` argument no longer needs to be specified.